// src/client/graphwalker/constants.ts
export enum ImageIdentifier {
  VertexIcon = "vertexIcon",
  Arrow = "arrow",
}

export enum Justification {
  Left = "left",
  Center = "center",
  Right = "right",
}

export enum VertexScrollDirection {
  Up = "up",
  Down = "down",
}

export type GraphicalIdentifier = ImageIdentifier;

export enum IncidentEdgeDirection {
  Incoming = "incoming",
  Outgoing = "outgoing",
}

export enum ScrollDirection {
  Up = "up",
  Down = "down",
}